requirements.txt
Django==4.0.6
django-extensions==3.1.5
djangorestframework==3.13.1
psycopg2-binary==2.9.3
requests-oauthlib==1.3.1
celery==5.2.7
redis==4.3.4
